Issue#4: Refactore dthe time type variables. Added some logic for the combat controller in order to perform actions and advance to the next combo, this is still work in progress.

+++ CombatSystem/DataStructures/TimeBasedActionConfiguraton.cs
@@ -9,11 +9,11 @@ namespace CombatSystem.DataStructures
 {
     public class TimeBasedActionConfiguraton : IActionConfiguraton
     {
-        public int PerformTime { get; private set; }
+        public float PerformTime { get; private set; }
 
         public AttackAction Action { get; set; }
 
-        public TimeBasedActionConfiguraton(int performTime)
+        public TimeBasedActionConfiguraton(float performTime)
         {
             PerformTime = performTime;
         }
